+++ R/sendrecv.R
@@ -22,11 +22,10 @@
 #'
 #' @param con a Socket, Context or Stream.
 #' @param data an object (a vector, if mode = \sQuote{raw}).
-#' @param mode [default 'serial'] character value or integer equivalent - one of
-#'     \sQuote{serial} (1L) to send serialised R objects, \sQuote{raw} (2L) to
-#'     send atomic vectors of any type as a raw byte vector, or \sQuote{next}
-#'     (3L) - see \sQuote{Send Modes} section below. For Streams, \sQuote{raw}
-#'     is the only option and this argument is ignored.
+#' @param mode [default 'serial'] character value or integer equivalent - either
+#'     \sQuote{serial} (1L) to send serialised R objects, or \sQuote{raw} (2L)
+#'     to send atomic vectors of any type as a raw byte vector. For Streams,
+#'     \sQuote{raw} is the only option and this argument is ignored.
 #' @param block [default NULL] which applies the connection default (see section
 #'     \sQuote{Blocking} below). Specify logical TRUE to block until successful
 #'     or FALSE to return immediately even if unsuccessful (e.g. if no
@@ -53,18 +52,14 @@
 #'
 #'     The default mode \sQuote{serial} sends serialised R objects to ensure
 #'     perfect reproducibility within R. When receiving, the corresponding mode
-#'     \sQuote{serial} should be used.
+#'     \sQuote{serial} should be used. Custom serialization and unserialization
+#'     functions for reference objects may be enabled by the function
+#'     \code{\link{next_config}}.
 #'
 #'     Mode \sQuote{raw} sends atomic vectors of any type as a raw byte vector,
 #'     and must be used when interfacing with external applications or raw
 #'     system sockets, where R serialization is not in use. When receiving, the
 #'     mode corresponding to the vector sent should be used.
-#'
-#'     Mode \sQuote{next} sends serialised R objects, with native extensions
-#'     enabled by \code{\link{next_config}}. This configures custom
-#'     serialization and unserialization functions for reference objects. When
-#'     receiving, mode \sQuote{serial} should be used as \sQuote{next} sends are
-#'     fully compatible.
 #'
 #' @seealso \code{\link{send_aio}} for asynchronous send.
 #' @examples
@@ -89,7 +84,7 @@
 #'
 #' @export
 #'
-send <- function(con, data, mode = c("serial", "raw", "next"), block = NULL)
+send <- function(con, data, mode = c("serial", "raw"), block = NULL)
   .Call(rnng_send, con, data, mode, block)
 
 #' Receive